modules renamed, const instead of let, html templates removed
