+++ js/main.js
@@ -1,6 +1,6 @@
-import * as dom from './create-dom-element';
-import moduleOneElement from './module-1';
+import * as dom from './dom-helpers';
+import welcomeElement from './template-welcome';
 
 window.onload = () => {
-  dom.renderElement(moduleOneElement);
+  dom.renderElement(welcomeElement);
 };
